requirements.txt
arabic_reshaper==3.0.0
Django==4.2.2
python_bidi==0.4.2
django-ckeditor==6.1.0
gunicorn
whitenoise==6.5.0